Added integer type checking in conditional path template expressions. Moved redundant condiitonalExpFormat into utils.py
