+++ tools/python/boutiques/tests/output_files/test_fixANDcond_output_desc.json
@@ -26,6 +26,7 @@
             "name": "option 2",
             "optional": true,
             "type": "Number",
+            "integer": true,
             "value-key": "[OPT2]"
         },
         {
@@ -60,7 +61,7 @@
             "name": "output 2",
             "optional": false,
             "conditional-path-template": [
-                {"opt1 and (opt2 >10.00)": "[PARAM1]_large.txt"},
+                {"opt1 and (opt2 > 9)": "[PARAM1]_large.txt"},
                 {"opt1":                   "[PARAM1].txt"},
                 {"opt2>10":                "[PARAM2]_large.txt"},
                 {"default":                "[PARAM2]_default.txt"}
